mudssky_utils

Module object_utils

Object utility functions similar to JavaScript object methods

This module provides utility functions for working with objects and data structures that are commonly available in JavaScript but not natively in Rust.

Enums

ObjectUtilsError

Error type for object operations

Functions

assign

Assign properties from source maps to target map Similar to JavaScript’s Object.assign()

deep_clone

Deep clone a nested HashMap structure Similar to lodash’s cloneDeep() for objects

entries

Get all key-value pairs from a HashMap Similar to JavaScript’s Object.entries()

from_entries

Create a HashMap from key-value pairs Similar to JavaScript’s Object.fromEntries()

has_key

Check if a HashMap has a specific key Similar to JavaScript’s Object.hasOwnProperty()

is_empty

Check if a HashMap is empty Similar to lodash’s isEmpty() for objects

keys

Get all keys from a HashMap Similar to JavaScript’s Object.keys()

merge

Merge multiple HashMaps into a new one Similar to JavaScript’s spread operator {…obj1, …obj2}

omit

Omit specific keys from a HashMap Similar to lodash’s omit() function

pick

Pick specific keys from a HashMap Similar to lodash’s pick() function

size

Get the size/length of a HashMap Similar to JavaScript’s Object.keys().length

values

Get all values from a HashMap Similar to JavaScript’s Object.values()